Random selector Explore View fix
