+++ menu/drivers/rgui.c
@@ -7989,6 +7989,7 @@ static enum menu_action rgui_parse_menu_entry_action(
       menu_entry_t *entry,
       enum menu_action action)
 {
+   struct menu_state *menu_st  = menu_state_get_ptr();
    enum menu_action new_action = action;
 
    /* Scan user inputs */
@@ -8056,7 +8057,6 @@ static enum menu_action rgui_parse_menu_entry_action(
 
          if (string_is_equal(rgui->menu_title, msg_hash_to_str(MENU_ENUM_LABEL_VALUE_MAIN_MENU)))
          {
-            struct menu_state *menu_st = menu_state_get_ptr();
             /* Jump to first item on Main Menu */
             menu_st->selection_ptr     = 0;
             new_action                 = MENU_ACTION_NOOP;
@@ -8112,7 +8112,6 @@ static enum menu_action rgui_parse_menu_entry_action(
       case MENU_ACTION_SCAN:
          if (rgui->flags & RGUI_FLAG_IS_PLAYLISTS_TAB)
          {
-            struct menu_state *menu_st = menu_state_get_ptr();
             size_t selection_total     = menu_st->entries.list ? MENU_LIST_GET_SELECTION(menu_st->entries.list, 0)->size : 0;
             size_t selection           = menu_st->selection_ptr;
             size_t new_selection       = random_range(0, selection_total - 1);
@@ -8145,29 +8144,8 @@ static enum menu_action rgui_parse_menu_entry_action(
          else if ((rgui->flags & RGUI_FLAG_IS_PLAYLIST)
                || (rgui->flags & RGUI_FLAG_IS_EXPLORE_LIST))
          {
-            struct menu_state *menu_st = menu_state_get_ptr();
-            size_t selection_start     = 0;
-            size_t selection_total     = menu_st->entries.list ? MENU_LIST_GET_SELECTION(menu_st->entries.list, 0)->size : 0;
-            size_t selection           = menu_st->selection_ptr;
-            size_t new_selection       = selection;
-
-            /* Skip header items (Search Name + Add Additional Filter + Save as View) */
-            if (rgui->flags & RGUI_FLAG_IS_EXPLORE_LIST)
-            {
-               menu_entry_t entry;
-               MENU_ENTRY_INITIALIZE(entry);
-               menu_entry_get(&entry, 0, 0, NULL, true);
-
-               if (entry.type == MENU_SETTINGS_LAST + 1)
-                  selection_start = 1;
-               else if (entry.type == FILE_TYPE_RDB)
-                  selection_start = 2;
-            }
-
-            new_selection = random_range(selection_start, selection_total - 1);
-
-            while (new_selection == selection && selection_start != selection_total - 1)
-               new_selection = random_range(selection_start, selection_total - 1);
+            size_t selection     = menu_st->selection_ptr;
+            size_t new_selection = menu_playlist_random_selection(selection, rgui->flags & RGUI_FLAG_IS_EXPLORE_LIST);
 
             if (new_selection != selection)
             {
